requirements.txt
torch
torchvision
numpy
opencv-python
tqdm
matplotlib
scikit-learn
scikit-image
scipy
h5py
pillow
loguru
pyopengl
pyopengl_accelerate
pyglm
trimesh
videoio
pycolmap
xmltodict
